Very primitive expression parsing for the SQL planner.

+++ src/net/eagledb/server/planner/PageScan.java
@@ -27,7 +27,7 @@ public class PageScan extends PageOperation {
 	public void run(TransactionPage tp, Page p, boolean[][] buf) {
 		if(p instanceof net.eagledb.server.storage.page.IntPage) {
 			IntPage page = (IntPage) p;
-			int compare = Integer.valueOf(value.toString());
+			int compare = (int) (double) Double.valueOf(value.toString());
 			if(action == PageAction.GREATER_THAN)
 				for(int i = 0; i < Page.TUPLES_PER_PAGE; ++i)
 					buf[buffer][i] = (page.page[i] > compare);
